[NO-JIRA] Make BpkModalV2 the default
