+++ packages/bpk-component-modal/index.ts
@@ -16,14 +16,6 @@
  * limitations under the License.
  */
 
-import { BpkModalV2 } from './src/BpkModalV2/BpkModal';
-import themeAttributes from './src/themeAttributes';
-import BpkModal from './src/BpkModal';
-import type { Props } from './src/BpkModal';
-// @ts-expect-error Untyped import. See `decisions/imports-ts-suppressions.md`.
-import { propTypes, defaultProps } from './src/legacy-prop-types';
-
-export type BpkModalProps = Props;
+import { BpkModal } from './src/BpkModal';
 
 export default BpkModal;
-export { propTypes, defaultProps, themeAttributes, BpkModalV2 };
