+++ packages/bpk-component-modal/src/BpkModal-test.tsx
@@ -16,31 +16,303 @@
  * limitations under the License.
  */
 
-import { render } from '@testing-library/react';
+import { render, fireEvent, screen } from '@testing-library/react';
+import '@testing-library/jest-dom';
 
-import BpkModal from './BpkModal';
+import { type BpkModal as DialogType } from './BpkModal';
+
+// this is needed so that we avoid importing multiple copies of React due to the isolateModules method
+// i.e. inside the BpkModal module as well as in the tests.
+// Importing multiple copies would result in an error, mocking it ensures tests are always running with the same React version
+jest.mock('react', () => jest.requireActual('react'));
 
 describe('BpkModal', () => {
-  it('should render correctly in the given target if renderTarget is supplied', () => {
-    const customRenderTarget = document.createElement('div');
+  const props = {
+    id: 'bpk-modal-element',
+    ariaLabelledby: 'bpk-modal-element',
+    closeLabel: 'bpk-modal-button-close',
+    isOpen: true,
+    onClose: jest.fn(),
+  };
 
-    const { asFragment } = render(
-      <BpkModal
-        id="my-modal"
-        title="Modal title"
-        onClose={jest.fn()}
-        closeLabel="Close"
-        dialogRef={jest.fn()}
-        isIphone={false}
-        getApplicationElement={jest.fn()}
-        isOpen
-        renderTarget={() => customRenderTarget}
-      >
-        Modal content inside a custom target
-      </BpkModal>,
-    );
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
 
-    expect(asFragment()).toMatchSnapshot();
-    expect(customRenderTarget).toMatchSnapshot();
+  describe('is supported', () => {
+    let BpkModal: typeof DialogType;
+
+    beforeEach(() => {
+      jest.isolateModules(() => {
+        ({ BpkModal } = jest.requireActual('./BpkModal'));
+      });
+    });
+
+    it('should render correctly with content', () => {
+      const { asFragment } = render(
+        <BpkModal {...props} title="Modal with dialog element">
+          <div>Content</div>
+        </BpkModal>,
+      );
+
+      expect(asFragment()).toMatchSnapshot();
+    });
+
+    it('should call on Close when closing the dialog', () => {
+      render(
+        <BpkModal {...props}>
+          <div>Content</div>
+        </BpkModal>,
+      );
+      fireEvent.click(screen.getByTitle('bpk-modal-button-close'));
+
+      expect(props.onClose).toHaveBeenCalledTimes(1);
+    });
+
+    it('should close the dialog when clicking outside the dialog', () => {
+      render(
+        <>
+          <div data-testid="page" />
+          <BpkModal {...props}>
+            <div>Content</div>
+          </BpkModal>
+        </>,
+      );
+      fireEvent.click(screen.getByTestId('page'));
+
+      expect(screen.queryByRole('Dialog')).toBeNull();
+    });
+
+    it('should not close the dialog when clicking inside the dialog', () => {
+      render(
+        <BpkModal {...props}>
+          <div>Content</div>
+        </BpkModal>,
+      );
+      fireEvent.click(screen.getByText('Content'));
+
+      expect(document.getElementById('bpk-modal-element')).toBeInTheDocument();
+    });
+
+    it('should render title correctly', () => {
+      render(
+        <BpkModal {...props} title="Modal with dialog element">
+          <div>Content</div>
+        </BpkModal>,
+      );
+
+      expect(screen.getByText('Modal with dialog element')).toBeInTheDocument();
+      expect(screen.getByTitle('bpk-modal-button-close')).toBeInTheDocument();
+    });
+
+    it('should not render title when title does not exist', () => {
+      render(
+        <BpkModal {...props}>
+          <div>Content</div>
+        </BpkModal>,
+      );
+
+      expect(
+        screen.queryByText('Modal with dialog element'),
+      ).not.toBeInTheDocument();
+      expect(screen.getByTitle('bpk-modal-button-close')).toBeInTheDocument();
+    });
+
+    it('should not render header when showHeader is false', () => {
+      render(
+        <BpkModal {...props} showHeader={false}>
+          <div>Content</div>
+        </BpkModal>,
+      );
+
+      expect(document.getElementById('bpk-modal-element')).toBeInTheDocument();
+      expect(document.getElementById('bpk-modal-element-title')).toBeNull();
+    });
+
+    it('should call showModal to open dialog', () => {
+      window.HTMLDialogElement.prototype.showModal = jest.fn();
+      render(
+        <BpkModal {...props}>
+          <div>Content</div>
+        </BpkModal>,
+      );
+
+      expect(
+        window.HTMLDialogElement.prototype.showModal,
+      ).toHaveBeenCalledTimes(1);
+    });
+
+    it('should return null when dialog is closed', () => {
+      // @ts-expect-error Untyped import. See `decisions/imports-ts-suppressions.md`.
+      const { rerender, store } = render(
+        <BpkModal {...props}>
+          <div>Content</div>
+        </BpkModal>,
+      );
+
+      expect(screen.queryByText('Content')).toBeInTheDocument();
+
+      rerender(
+        <BpkModal {...props} isOpen={false}>
+          <div>Content</div>
+        </BpkModal>,
+        // @ts-expect-error Untyped import. See `decisions/imports-ts-suppressions.md`.
+        rerender,
+        store,
+      );
+
+      expect(screen.queryByRole('Dialog')).toBeNull();
+    });
+
+    it('should set up the correct className when fullScreenOnDesktop is true', () => {
+      render(
+        <BpkModal {...props} fullScreenOnDesktop>
+          <div>Content</div>
+        </BpkModal>,
+      );
+
+      expect(document.getElementById('bpk-modal-element')).toHaveClass(
+        'bpk-modal bpk-modal--full-screen-desktop',
+      );
+      expect(
+        document.getElementsByClassName(
+          'bpk-modal_container bpk-modal__container--full-screen-desktop',
+        ),
+      ).toBeTruthy();
+    });
+
+    it('should set up the correct classNames when noFullScreenOnMobile is true', () => {
+      render(
+        <BpkModal {...props} noFullScreenOnMobile>
+          <div>Content</div>
+        </BpkModal>,
+      );
+
+      expect(document.getElementById('bpk-modal-element')).toHaveClass(
+        'bpk-modal bpk-modal--no-full-screen-mobile',
+      );
+    });
+
+    it('should set up the correct className when padded is false', () => {
+      render(
+        <BpkModal {...props} padded={false}>
+          <div>Content</div>
+        </BpkModal>,
+      );
+
+      expect(
+        document.getElementsByClassName(
+          'bpk-modal_container bpk-modal__container--padded',
+        ),
+      ).toBeTruthy();
+    });
+
+    it('should set up the correct className when wide is true', () => {
+      render(
+        <BpkModal {...props} wide>
+          <div>Content</div>
+        </BpkModal>,
+      );
+
+      expect(document.getElementById('bpk-modal-element')).toHaveClass(
+        'bpk-modal bpk-modal--wide',
+      );
+    });
+
+    describe('setOverflowY', () => {
+      it.each([
+        ['hidden', true],
+        ['visible', false],
+      ])('to %p when isOpen is %p', (expectedOverflowY, expectedIsOpen) => {
+        render(
+          <BpkModal {...props} isOpen={expectedIsOpen}>
+            <div>Content</div>
+          </BpkModal>,
+        );
+
+        expect(document.body.style.overflowY).toEqual(expectedOverflowY);
+      });
+    });
+  });
+
+  describe('is not supported', () => {
+    let htmlDialogElement: typeof window.HTMLDialogElement;
+
+    let BpkModal: typeof DialogType;
+
+    beforeEach(async () => {
+      htmlDialogElement = window.HTMLDialogElement;
+      window.HTMLDialogElement = undefined!;
+
+      jest.isolateModules(() => {
+        ({ BpkModal } = jest.requireActual('./BpkModal'));
+      });
+    });
+
+    afterEach(() => {
+      window.HTMLDialogElement = htmlDialogElement;
+    });
+
+    it('should render correctly with polyfill and content', () => {
+      const { asFragment } = render(
+        <BpkModal {...props}>
+          <div>Content</div>
+        </BpkModal>,
+      );
+
+      expect(asFragment()).toMatchSnapshot();
+    });
+
+    it('should call use the polyfill to open the dialog', () => {
+      render(
+        <BpkModal {...props}>
+          <div>Content</div>
+        </BpkModal>,
+      );
+
+      expect(
+        document
+          .getElementById('bpk-modal-element-polyfill')
+          ?.getAttribute('data-open'),
+      ).toEqual('true');
+      expect(document.body.style.position).toEqual('fixed');
+      expect(document.body.style.width).toEqual('100%');
+    });
+
+    it('should call on Close when closing the dialog', () => {
+      render(
+        <BpkModal {...props}>
+          <div>Content</div>
+        </BpkModal>,
+      );
+      fireEvent.click(screen.getByTitle('bpk-modal-button-close'));
+
+      expect(props.onClose).toHaveBeenCalledTimes(1);
+    });
+
+    it('should close the dialog when clicking outside the dialog', () => {
+      render(
+        <>
+          <div data-testid="page" />
+          <BpkModal {...props}>
+            <div>Content</div>
+          </BpkModal>
+        </>,
+      );
+      fireEvent.click(screen.getByTestId('page'));
+
+      expect(screen.queryByRole('Dialog')).toBeNull();
+    });
+
+    it('should reset position and width when dialog is closed', () => {
+      render(
+        <BpkModal {...props} isOpen={false}>
+          <div>Content</div>
+        </BpkModal>,
+      );
+
+      expect(document.body.style.position).toEqual('relative');
+      expect(document.body.style.width).toEqual('auto');
+    });
   });
 });
