+++ packages/bpk-component-modal/src/BpkModal.module.scss
@@ -16,31 +16,165 @@
  * limitations under the License.
  */
 
-@import '../../bpk-mixins/index.scss';
+@import '../../../bpk-mixins/index.scss';
 
 @layer bpk-component {
-  @mixin fullscreen() {
+  @mixin header {
     display: flex;
-    padding: 0;
+    min-height: bpk-spacing-lg();
+    flex-wrap: nowrap;
+    justify-content: flex-end;
+    align-items: center;
   }
 
-  .bpk-modal__container {
-    display: flex;
-    padding: bpk-spacing-lg();
-
-    // IE11 and below hack for buggy flexbox support
-    @media screen\0 {
-      display: block;
+  @keyframes bpk-modal-scrim {
+    0% {
+      opacity: $bpk-modal-initial-opacity;
     }
 
-    &--full-screen-mobile {
-      @include bpk-breakpoint-mobile {
-        @include fullscreen;
+    100% {
+      opacity: $bpk-modal-opacity;
+    }
+  }
+
+  @keyframes bpk-modal-scale {
+    0% {
+      transform: scale(0.9);
+    }
+
+    100% {
+      transform: scale(1);
+    }
+  }
+
+  .bpk-modal-wrapper.bpk-modal-polyfill {
+    .bpk-modal-backdrop {
+      position: fixed;
+      z-index: 0;
+      margin: 0 auto;
+      background-color: $bpk-scrim-day;
+      animation: bpk-modal-scrim $bpk-duration-sm ease-in-out;
+      inset: 0;
+      inset-block-end: 0;
+      inset-inline: 0;
+      inset-inline-end: 0;
+    }
+
+    .bpk-modal {
+      position: fixed;
+      top: 50%;
+      left: 50%;
+      display: block;
+      z-index: 1;
+      transform: translate(-50%, -50%) scale(1);
+      transition: transform $bpk-duration-sm ease-in-out $bpk-duration-sm;
+    }
+
+    .bpk-modal:not([data-open='true']),
+    .bpk-modal-backdrop:not([data-open='true']) {
+      display: none;
+    }
+  }
+
+  .bpk-modal {
+    width: 100%;
+    max-width: none;
+    height: 100%;
+    max-height: 100%;
+    padding: 0;
+    border: none;
+    background: $bpk-modal-background-color;
+    overflow-y: scroll;
+    scrollbar-width: none;
+
+    @include bpk-box-shadow-xl;
+
+    @include bpk-breakpoint-above-mobile {
+      width: $bpk-modal-max-width;
+      max-width: none;
+      height: fit-content;
+      max-height: 90%;
+      border-radius: $bpk-border-radius-md;
+
+      &--full-screen-desktop {
+        width: 100%;
+        height: 100%;
+        max-height: 100%;
+        border-radius: 0;
+      }
+
+      &--wide {
+        width: $bpk-modal-wide-max-width;
+        max-width: 100%;
       }
     }
 
-    &--full-screen {
-      @include fullscreen;
+    &[open] {
+      animation: bpk-modal-scale $bpk-duration-sm ease-in-out;
+    }
+
+    &--no-full-screen-mobile {
+      width: 90%;
+      max-width: none;
+      height: fit-content;
+      max-height: 90%;
+
+      @include bpk-breakpoint-above-mobile {
+        width: $bpk-modal-max-width;
+        max-width: none;
+      }
+    }
+
+    &::backdrop {
+      position: fixed;
+      background-color: $bpk-scrim-day;
+      animation: bpk-modal-scrim $bpk-duration-sm ease-in-out;
+      inset: 0;
+      inset-block-end: 0;
+      inset-inline: 0;
+      inset-inline-end: 0;
+    }
+
+    &::-webkit-scrollbar {
+      display: none;
+    }
+
+    &__header-title {
+      padding: bpk-spacing-base();
+
+      @include header;
+    }
+
+    &__header-title-container {
+      display: flex;
+      width: 100%;
+      justify-content: flex-start;
+    }
+
+    &__title {
+      padding-inline-end: bpk-spacing-base();
+
+      @include bpk-label-1;
+    }
+
+    &__button-container {
+      padding-block-start: bpk-spacing-base();
+      padding-inline-end: bpk-spacing-base();
+
+      @include header;
+    }
+
+    &__container {
+      min-height: 100%;
+      padding: $bpk-modal-content-padding;
+
+      &--padded {
+        padding: 0;
+      }
+
+      &--full-screen-desktop {
+        height: 100%;
+      }
     }
   }
 }
