--- packages/bpk-component-modal/src/BpkModalV2/BpkModal.d.ts
@@ -1,36 +1,0 @@
-/*
- * Backpack - Skyscanner's Design System
- *
- * Copyright 2016 Skyscanner Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/// <reference types="react" />
-import type { ReactNode } from 'react';
-
-export type Props = {
-  id: string | undefined;
-  ariaLabelledby: string;
-  children: ReactNode;
-  closeLabel: string;
-  fullScreenOnDesktop?: boolean;
-  isOpen: boolean;
-  noFullScreenOnMobile?: boolean;
-  onClose: () => void | null;
-  padded?: boolean;
-  showHeader?: boolean;
-  title?: string | null;
-  wide?: boolean;
-};
-export declare const BpkModalV2: (props: Props) => JSX.Element | null;
