+++ packages/bpk-component-modal/src/accessibility-test.tsx
@@ -19,26 +19,22 @@
 import { render } from '@testing-library/react';
 import { axe } from 'jest-axe';
 
-import BpkModal from './BpkModal';
+import { BpkModalV2 } from './BpkModal';
 
 describe('BpkModal accessibility tests', () => {
   it('should not have programmatically-detectable accessibility issues', async () => {
-    const customRenderTarget = document.createElement('div');
-
     const { container } = render(
-      <BpkModal
-        id="my-modal"
-        title="Modal title"
-        onClose={jest.fn()}
-        closeLabel="Close"
-        dialogRef={jest.fn()}
-        isIphone={false}
-        getApplicationElement={jest.fn()}
+      <BpkModalV2
+        id="bpk-modal-element"
+        ariaLabelledby="bpk-modal-label-my-dialog"
+        closeLabel="bpk-modal-button-close"
         isOpen
-        renderTarget={() => customRenderTarget}
+        onClose={jest.fn()}
+        title="Backpack Dialog Element"
+        padded
       >
-        Modal content inside a custom target
-      </BpkModal>,
+        Dialog content
+      </BpkModalV2>,
     );
     const results = await axe(container);
     expect(results).toHaveNoViolations();
